backend/requirements.txt
asgiref
attrs
awscli
behave
boto3
certifi
charset-normalizer
coverage
Django
django-cors-headers
django-environ
django-extensions
django-storages
djangorestframework
djangorestframework_simplejwt
h11
idna
iniconfig
outcome
packaging
parse
parse_type
pluggy
psycopg2-binary
PyJWT
PySocks
pytest
pytest-django
pytest-mock
python-dotenv
requests
selenium
six
sniffio
sortedcontainers
sqlparse
trio
trio-websocket
typing_extensions
urllib3
webdriver-manager
websocket-client
wheel
wsproto
